+++ src/main/java/com/demo/thymeleaf/repository/CinRepostiroy.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface CinRepostiroy extends JpaRepository<Cin, Integer> {
     @Query(value = "SELECT * FROM cin WHERE number = ?1 ", nativeQuery = true)
-    Cin get_cin_byNumber(Double number);
+    Cin get_cin_byNumber(Integer number);
 }
